add validation service to interpolate heart rate measurement data

+++ lib/session/data/model/session_parameter_optimization.dart
@@ -1,5 +1,5 @@
 class SessionParameterOptimization {
-  final double beatFrequency;
+  final int beatFrequency;
   final String visualization;
   final double breathingPatternMultiplier;
